将所有doule类型修改位float类型

+++ determinant.cpp
@@ -1,6 +1,6 @@
 #include "determinant.h"
 #include <iostream>
-#include <qvector.h>
+#include <QVector>
 #include <QDebug>
 using namespace std;
 Determinant::Determinant()
@@ -68,21 +68,21 @@ bool Determinant::PowerIsPosition(QVector<int> &vec)
     return (Iseven(count));
 }
 
-double Determinant::calculate(QVector<QVector<double>> vec_Determinant, int n)
+float Determinant::calculate(QVector<QVector<float>> vec_Determinant, int n)
 {
     QVector<QVector<int>> vec_que;
     QVector<int> vec = inivec(n);
     QVector<int> vec_elem;
     //    最终结果初始化为0
     Perm(vec, &vec_que,0,vec.size()-1);
-    double result = 0;
+    float result = 0;
     //    依次从vec_que中取出行列式
     for (int i = 0; i < vec_que.size(); i++)
     {
         vec_elem = vec_que[i];
         //    mi即为前面(-1)的n次幂，最后结果为-1或者1
         int mi = PowerIsPosition(vec_elem) ? 1 : -1;
-        double temp  = mi;
+        float temp  = mi;
         //        row号初始化为0之后依次加1
         int row = 0;
         //        col号依次从vec_elem中取出
